+++ lamindb/models/query_set.py
@@ -1513,11 +1513,12 @@ class QuerySet(BasicQuerySet):
             raise  # pragma: no cover
 
         if is_run_input is not False:  # might be None or True or Run
-            from .artifact import Artifact, track_run_input
+            from ._track_run_inputs import track_run_inputs
+            from .artifact import Artifact
             from .collection import Collection
 
             if isinstance(record, (Artifact, Collection)):
-                track_run_input(record, is_run_input)
+                track_run_inputs(record, is_run_input)
 
         return record
