+++ tests/permissions/test_rls_dbwritelog.py
@@ -11,7 +11,7 @@ from django.db import connection, transaction
 from django.db.utils import IntegrityError, InternalError, ProgrammingError
 from hubmodule.sql_generators._dbwrite import uninstall_dbwrite
 from jwt_utils import sign_jwt
-from lamindb.models.artifact import track_run_input
+from lamindb.models._track_run_inputs import track_run_inputs
 from lamindb_setup.core.django import DBToken, db_token_manager
 from psycopg2.extensions import adapt
 
@@ -461,20 +461,20 @@ def test_tracking_error():
     # this error because ln.setup.settings.instance._db_permissions is not jwt
     # it is None
     with pytest.raises(ln.errors.NoWriteAccess) as e:
-        track_run_input(artifact, run)
+        track_run_inputs(artifact, run)
     assert "You’re not allowed to write to the instance " in str(e)
 
     # the instance is local so we set this manually
     ln.setup.settings.instance._db_permissions = "jwt"
     # artifact.space is not available for writes
     with pytest.raises(ln.errors.NoWriteAccess) as e:
-        track_run_input(artifact, run)
+        track_run_inputs(artifact, run)
     assert "You’re not allowed to write to the space " in str(e)
 
     # this artifact is locked
     artifact = ln.Artifact.get(description="test locking")
     with pytest.raises(ln.errors.NoWriteAccess) as e:
-        track_run_input(artifact, run)
+        track_run_inputs(artifact, run)
     assert "It is not allowed to modify locked sqlrecords" in str(e)
 
     # switch user role back to read
@@ -485,7 +485,7 @@ def test_tracking_error():
     # as the user is read-only now, 2 spaces are unavailable for writes (artifact.space, run.space)
     artifact = ln.Artifact.get(description="test tracking error")
     with pytest.raises(ln.errors.NoWriteAccess) as e:
-        track_run_input(artifact, run)
+        track_run_inputs(artifact, run)
     assert "You’re not allowed to write to the spaces " in str(e)
 
     ln.setup.settings.instance._db_permissions = None
